# backend_openapi.json
{"openapi":"3.1.0","info":{"title":"Web Intelligence Ingestion Engine","description":"Production-grade brand intelligence data foundation","version":"1.0.0"},"paths":{"/health/":{"get":{"tags":["Health"],"summary":"Health","operationId":"health_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/search/":{"post":{"tags":["Search"],"summary":"Submit Search","description":"v3: Single batch INSERT for all URLs (vs N sequential inserts in v2).\nEnriches url_hash on results list in-memory before enqueue.","operationId":"submit_search_api_v1_search__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/search/status/{job_id}":{"get":{"tags":["Search"],"summary":"Job Status","operationId":"job_status_api_v1_search_status__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/crawl/queue":{"get":{"tags":["Crawl"],"summary":"Queue Stats","operationId":"queue_stats_api_v1_crawl_queue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/content/":{"get":{"tags":["Content"],"summary":"Get Content","description":"Retrieve processed content for a query.\nReturns documents matching the canonical v2 schema.","operationId":"get_content_api_v1_content__get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","minLength":2,"title":"Query"}},{"name":"min_relevance","in":"query","required":false,"schema":{"type":"number","maximum":1.0,"minimum":0.0,"default":0.4,"title":"Min Relevance"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"sort_by","in":"query","required":false,"schema":{"type":"string","default":"relevance_score","title":"Sort By"}},{"name":"language","in":"query","required":false,"schema":{"type":"string","title":"Language"}},{"name":"credibility","in":"query","required":false,"schema":{"type":"string","title":"Credibility"}},{"name":"include_text","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Text"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/content/search":{"get":{"tags":["Content"],"summary":"Fulltext Search","description":"Full-text search across content_summary, relevant_text, key_points, title.","operationId":"fulltext_search_api_v1_content_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":3,"description":"Full-text search term","title":"Q"},"description":"Full-text search term"},{"name":"query","in":"query","required":false,"schema":{"type":"string","description":"Filter by original search query","title":"Query"},"description":"Filter by original search query"},{"name":"min_relevance","in":"query","required":false,"schema":{"type":"number","maximum":1.0,"minimum":0.0,"default":0.4,"title":"Min Relevance"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/content/{url_hash}":{"get":{"tags":["Content"],"summary":"Get By Hash","description":"Get a specific document by URL hash. full=true includes clean_text and embedding_text.","operationId":"get_by_hash_api_v1_content__url_hash__get","parameters":[{"name":"url_hash","in":"path","required":true,"schema":{"type":"string","title":"Url Hash"}},{"name":"full","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Full"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/content/domain/{domain}":{"get":{"tags":["Content"],"summary":"Get By Domain","description":"Get all content from a specific source domain.","operationId":"get_by_domain_api_v1_content_domain__domain__get","parameters":[{"name":"domain","in":"path","required":true,"schema":{"type":"string","title":"Domain"}},{"name":"query","in":"query","required":false,"schema":{"type":"string","title":"Query"}},{"name":"min_relevance","in":"query","required":false,"schema":{"type":"number","default":0.4,"title":"Min Relevance"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/summary":{"get":{"tags":["Analytics"],"summary":"Analytics Summary","operationId":"analytics_summary_api_v1_analytics_summary_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","minLength":2,"title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/domains":{"get":{"tags":["Analytics"],"summary":"Domain Breakdown","operationId":"domain_breakdown_api_v1_analytics_domains_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/quality":{"get":{"tags":["Analytics"],"summary":"Quality Distribution","description":"Relevance score distribution for a query — useful for threshold tuning.","operationId":"quality_distribution_api_v1_analytics_quality_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobStatus":{"properties":{"job_id":{"type":"string","title":"Job Id"},"query":{"type":"string","title":"Query"},"total_urls":{"type":"integer","title":"Total Urls"},"pending":{"type":"integer","title":"Pending"},"processing":{"type":"integer","title":"Processing"},"completed":{"type":"integer","title":"Completed"},"failed":{"type":"integer","title":"Failed"},"skipped":{"type":"integer","title":"Skipped"},"progress_pct":{"type":"number","title":"Progress Pct"}},"type":"object","required":["job_id","query","total_urls","pending","processing","completed","failed","skipped","progress_pct"],"title":"JobStatus"},"SearchRequest":{"properties":{"query":{"type":"string","maxLength":500,"minLength":2,"title":"Query"},"max_results":{"type":"integer","maximum":50.0,"minimum":1.0,"title":"Max Results","default":30},"search_engine":{"type":"string","title":"Search Engine","default":"auto"}},"type":"object","required":["query"],"title":"SearchRequest"},"SearchResponse":{"properties":{"job_id":{"type":"string","title":"Job Id"},"query":{"type":"string","title":"Query"},"urls_found":{"type":"integer","title":"Urls Found"},"queued":{"type":"integer","title":"Queued"},"message":{"type":"string","title":"Message"}},"type":"object","required":["job_id","query","urls_found","queued","message"],"title":"SearchResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}